Provide interpolated points to dlf
@sgkang , that is one of the resulting outcomes of our discussion today. It is now possible to provide the interpolated points to the `dlf`, so you save one call to `get_spline_values` (the one in `empymod.dlf`, not here).

However, this is only available in the dev-branch of empymod, so you better wait a bit with this pull request until the next version of empymod is released.

This would have the biggest impact if you loop over frequencies, specifically in time-domain calculations, where you need a lot of frequencies, but the offsets and therefore `lambd` and `int_pts` stay the same. But I think you do all frequencies at once, so the impact might actually be almost nil. Still thought it might be worth it.